Fix circular imports, fix bookkeeping interface
